refactor: rearrange modules
